+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.0
 - Core refactoring.
 - Added work time.
+- Pause, restart, stop.
 
 ## 1.0.1